Time: 268 ms (73.97%), Space: 132.7 MB (57.93%) - LeetHub

+++ 2379-maximum-total-importance-of-roads/2379-maximum-total-importance-of-roads.cpp
@@ -7,7 +7,7 @@ public:
             vec[it[1]]++;
         }
         long long ans = 0;
-        long long val = 1;
+        long val = 1;
         sort(vec.begin(),vec.end());
         for(auto it: vec) ans += it*(val++);
         return ans;
